+++ egs2/TEMPLATE/sds1/app.py
@@ -4,6 +4,16 @@ import shutil
 
 import torch
 from pyscripts.utils.dialog_eval.ASR_WER import handle_espnet_ASR_WER
+from pyscripts.utils.dialog_eval.human_feedback import (
+    natural_vote1_last_response,
+    natural_vote2_last_response,
+    natural_vote3_last_response,
+    natural_vote4_last_response,
+    relevant_vote1_last_response,
+    relevant_vote2_last_response,
+    relevant_vote3_last_response,
+    relevant_vote4_last_response,
+)
 from pyscripts.utils.dialog_eval.LLM_Metrics import (
     DialoGPT_perplexity,
     bert_score,
@@ -217,7 +227,6 @@ start_warmup()
 callback = gr.CSVLogger()
 start_record_time = None
 enable_btn = gr.Button(interactive=True, visible=True)
-disable_btn = gr.Button(interactive=False, visible=False)
 
 
 def flash_buttons():
@@ -229,91 +238,6 @@ def flash_buttons():
     ) + btn_updates
 
 
-def get_ip(request: gr.Request):
-    if "cf-connecting-ip" in request.headers:
-        ip = request.headers["cf-connecting-ip"]
-    elif "x-forwarded-for" in request.headers:
-        ip = request.headers["x-forwarded-for"]
-        if "," in ip:
-            ip = ip.split(",")[0]
-    else:
-        ip = request.client.host
-    return ip
-
-
-def natural_vote1_last_response(request: gr.Request):
-    ip_address1 = get_ip(request)
-    print(f"Very Natural (voted). ip: {ip_address1}")
-    return (
-        "Very Natural",
-        ip_address1,
-    ) + (disable_btn,) * 4
-
-
-def natural_vote2_last_response(request: gr.Request):
-    ip_address1 = get_ip(request)
-    print(f"Somewhat Awkward (voted). ip: {ip_address1}")
-    return (
-        "Somewhat Awkward",
-        ip_address1,
-    ) + (disable_btn,) * 4
-
-
-def natural_vote3_last_response(request: gr.Request):
-    ip_address1 = get_ip(request)
-    print(f"Very Awkward (voted). ip: {ip_address1}")
-    return (
-        "Very Awkward",
-        ip_address1,
-    ) + (disable_btn,) * 4
-
-
-def natural_vote4_last_response(request: gr.Request):
-    ip_address1 = get_ip(request)
-    print(f"Unnatural (voted). ip: {ip_address1}")
-    return (
-        "Unnatural",
-        ip_address1,
-    ) + (disable_btn,) * 4
-
-
-def relevant_vote1_last_response(request: gr.Request):
-    ip_address1 = get_ip(request)
-    print(f"Highly Relevant (voted). ip: {ip_address1}")
-    return (
-        "Highly Relevant",
-        ip_address1,
-    ) + (disable_btn,) * 4
-
-
-def relevant_vote2_last_response(request: gr.Request):
-    ip_address1 = get_ip(request)
-    print(f"Partially Relevant (voted). ip: {ip_address1}")
-    return (
-        "Partially Relevant",
-        ip_address1,
-    ) + (disable_btn,) * 4
-
-
-def relevant_vote3_last_response(request: gr.Request):
-    ip_address1 = get_ip(request)
-    print(f"Slightly Irrelevant (voted). ip: {ip_address1}")
-    return (
-        "Slightly Irrelevant",
-        ip_address1,
-    ) + (disable_btn,) * 4
-
-
-def relevant_vote4_last_response(request: gr.Request):
-    ip_address1 = get_ip(request)
-    print(f"Completely Irrelevant (voted). ip: {ip_address1}")
-    return (
-        "Completely Irrelevant",
-        ip_address1,
-    ) + (disable_btn,) * 4
-
-
-import json
 import time
