+++ egs2/voice_assistant_demo/sds1/pyscripts
@@ -1,0 +1,1 @@
+../../TEMPLATE/sds1/pyscripts/
